Separated out connection recovery tests
This will allow these slow tests to run in parallel
