+++ projects/Test/Integration/ConnectionRecovery/TestConsumerRecovery.cs
@@ -1,0 +1,85 @@
+// This source code is dual-licensed under the Apache License, version
+// 2.0, and the Mozilla Public License, version 2.0.
+//
+// The APL v2.0:
+//
+//---------------------------------------------------------------------------
+//   Copyright (c) 2007-2020 VMware, Inc.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       https://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+//---------------------------------------------------------------------------
+//
+// The MPL v2.0:
+//
+//---------------------------------------------------------------------------
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+//
+//  Copyright (c) 2007-2020 VMware, Inc.  All rights reserved.
+//---------------------------------------------------------------------------
+
+using System.Threading.Tasks;
+using RabbitMQ.Client;
+using RabbitMQ.Client.Events;
+using RabbitMQ.Client.Framing.Impl;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Test.Integration.ConnectionRecovery
+{
+    public class TestConsumerRecovery : TestConnectionRecoveryBase
+    {
+        public TestConsumerRecovery(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        [Fact]
+        public async Task TestConsumerRecoveryWithManyConsumers()
+        {
+            string q = (await _channel.QueueDeclareAsync(GenerateQueueName(), false, false, false)).QueueName;
+            int n = 1024;
+
+            for (int i = 0; i < n; i++)
+            {
+                var cons = new AsyncEventingBasicConsumer(_channel);
+                await _channel.BasicConsumeAsync(q, true, cons);
+            }
+
+            var tcs = new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
+            ((AutorecoveringConnection)_conn).ConsumerTagChangeAfterRecovery += (prev, current) => tcs.SetResult(true);
+
+            await CloseAndWaitForRecoveryAsync();
+            await WaitAsync(tcs, "consumer tag change after recovery");
+            Assert.True(_channel.IsOpen);
+            await AssertConsumerCountAsync(q, n);
+        }
+
+        [Fact]
+        public async Task TestThatCancelledConsumerDoesNotReappearOnRecovery()
+        {
+            string q = (await _channel.QueueDeclareAsync(GenerateQueueName(), false, false, false)).QueueName;
+            int n = 1024;
+
+            for (int i = 0; i < n; i++)
+            {
+                var cons = new AsyncEventingBasicConsumer(_channel);
+                string tag = await _channel.BasicConsumeAsync(q, true, cons);
+                await _channel.BasicCancelAsync(tag);
+            }
+            await CloseAndWaitForRecoveryAsync();
+            Assert.True(_channel.IsOpen);
+            await AssertConsumerCountAsync(q, 0);
+        }
+    }
+}
